Renamed RecordingtravelRVAdapter to TravelRecordRVAdapter  - 기록할 여행, 다녀온 여행 화면 둘다 똑같은 item ui를 사용하므로 하나의 어댑터로 관리
