+++ app/src/main/java/com/nexters/travelbudget/ui/main/record/RecordedTravelFragment.kt
@@ -9,9 +9,8 @@ import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
 import com.nexters.travelbudget.R
 import com.nexters.travelbudget.databinding.FragmentRecordedTravelBinding
 import com.nexters.travelbudget.ui.base.BaseFragment
-import com.nexters.travelbudget.ui.main.MainViewModel
+import com.nexters.travelbudget.ui.main.record.adapter.TravelRecordRVAdapter
 import com.nexters.travelbudget.utils.CustomItemDecoration
-import org.koin.androidx.viewmodel.ext.android.sharedViewModel
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
 /**
@@ -51,10 +50,11 @@ class RecordedTravelFragment :
 
     private fun setRecordedTravelRV() {
         binding.rvRecordedTravel.run {
-            adapter = RecordingTravelRVAdapter { tripRecordResponse ->
-                // TODO 여행 상세 화면으로 연결 작업
-                Toast.makeText(context, "상세 화면 전환", Toast.LENGTH_LONG).show()
-            }
+            adapter =
+                TravelRecordRVAdapter { tripRecordResponse ->
+                    // TODO 여행 상세 화면으로 연결 작업
+                    Toast.makeText(context, "상세 화면 전환", Toast.LENGTH_LONG).show()
+                }
             addItemDecoration(object : CustomItemDecoration() {
                 override fun setSpacingForDirection(
                     outRect: Rect,
